Move entri into own directory
