+++ apps/builder/app/shared/context.server.ts
@@ -2,6 +2,7 @@ import type { AppContext } from "@webstudio-is/trpc-interface/index.server";
 import env from "~/env/env.server";
 import { authenticator } from "~/services/auth.server";
 import { trpcClient } from "~/services/trpc.server";
+import { entryApi } from "./entri/entri-api.server";
 
 const createAuthorizationContext = async (
   request: Request
@@ -50,10 +51,7 @@ const createDeploymentContext = (request: Request) => {
 
 const createEntriContext = () => {
   return {
-    env: {
-      ENTRI_APPLICATION_ID: env.ENTRI_APPLICATION_ID,
-      ENTRI_SECRET: env.ENTRI_SECRET,
-    },
+    entryApi,
   };
 };
